# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "GameCardRead": {
        "additionalProperties": false,
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "critic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "developer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Developer"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "letsplay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LetsPlayRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "metacritic_slug": {
            "title": "Metacritic Slug",
            "type": "string"
          },
          "platforms": {
            "items": {
              "$ref": "#/components/schemas/PlatformScore"
            },
            "title": "Platforms",
            "type": "array"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "release_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "similar": {
            "items": {
              "$ref": "#/components/schemas/SimilarGameRef"
            },
            "title": "Similar",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "video_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Url"
          }
        },
        "required": [
          "metacritic_slug",
          "title"
        ],
        "title": "GameCardRead",
        "type": "object"
      },
      "GameListItemRead": {
        "additionalProperties": false,
        "properties": {
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "developer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Developer"
          },
          "metacritic_slug": {
            "title": "Metacritic Slug",
            "type": "string"
          },
          "metascore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metascore"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "title": "Platforms",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "userscore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userscore"
          }
        },
        "required": [
          "metacritic_slug",
          "title",
          "updated_at"
        ],
        "title": "GameListItemRead",
        "type": "object"
      },
      "GameListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameListItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "GameListResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonitorSnapshot": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "items": {
              "$ref": "#/components/schemas/StageStatusCountRead"
            },
            "title": "Counts",
            "type": "array"
          },
          "cursor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MonitorCursorRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MonitorItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "process_date": {
            "format": "date",
            "title": "Process Date",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/MonitorRunRead"
            },
            "title": "Runs",
            "type": "array"
          },
          "scrape": {
            "$ref": "#/components/schemas/MonitorScrapeRead"
          },
          "workers": {
            "items": {
              "$ref": "#/components/schemas/WorkerHeartbeatRead"
            },
            "title": "Workers",
            "type": "array"
          }
        },
        "required": [
          "process_date",
          "workers",
          "runs",
          "items",
          "counts",
          "scrape"
        ],
        "title": "MonitorSnapshot",
        "type": "object"
      },
      "PaginationMeta": {
        "additionalProperties": false,
        "properties": {
          "order": {
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Order",
            "type": "string"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "sort": {
            "enum": [
              "metascore",
              "userscore",
              "title",
              "updated"
            ],
            "title": "Sort",
            "type": "string"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total",
          "sort",
          "order"
        ],
        "title": "PaginationMeta",
        "type": "object"
      },
      "PlatformListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PlatformListResponse",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "ReviewSummary": {
        "additionalProperties": false,
        "properties": {
          "dislikes": {
            "items": {
              "type": "string"
            },
            "title": "Dislikes",
            "type": "array"
          },
          "likes": {
            "items": {
              "type": "string"
            },
            "title": "Likes",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "likes",
          "dislikes",
          "summary"
        ],
        "title": "ReviewSummary",
        "type": "object"
      },
      "RunAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "process_date": {
            "format": "date",
            "title": "Process Date",
            "type": "string"
          },
          "status": {
            "const": "run_accepted",
            "default": "run_accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "process_date"
        ],
        "title": "RunAcceptedResponse",
        "type": "object"
      },
      "SimilarGameRef": {
        "additionalProperties": false,
        "properties": {
          "metacritic_slug": {
            "title": "Metacritic Slug",
            "type": "string"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "metacritic_slug",
          "title",
          "score",
          "rank"
        ],
        "title": "SimilarGameRef",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "games-intel-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/games": {
      "get": {
        "operationId": "list_games_api_v1_games_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "metascore",
              "enum": [
                "metascore",
                "userscore",
                "title",
                "updated"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Games",
        "tags": [
          "games"
        ]
      }
    },
    "/api/v1/games/{slug}": {
      "get": {
        "operationId": "get_game_api_v1_games__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameCardRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Game",
        "tags": [
          "games"
        ]
      }
    },
    "/api/v1/media/covers/{slug}": {
      "get": {
        "operationId": "get_cover_api_v1_media_covers__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {},
              "image/png": {},
              "image/webp": {}
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "description": "Cover file is missing"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Cover",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/monitor": {
      "get": {
        "operationId": "get_monitor_api_v1_monitor_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Monitor",
        "tags": [
          "monitor"
        ]
      }
    },
    "/api/v1/monitor/stream": {
      "get": {
        "operationId": "stream_monitor_api_v1_monitor_stream_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE кадры снимка монитора"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Stream Monitor",
        "tags": [
          "monitor"
        ]
      }
    },
    "/api/v1/platforms": {
      "get": {
        "operationId": "list_platforms_api_v1_platforms_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Platforms",
        "tags": [
          "games",
          "platforms"
        ]
      }
    },
    "/api/v1/runs": {
      "post": {
        "operationId": "start_run_api_v1_runs_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Run",
        "tags": [
          "monitor"
        ]
      }
    },
...
  }
}